Minor refacgtor for routeHandlers.
